add django check and dependency on modeladmin
The UI for wagtail-sharing depends on wagtail.contrib.modeladmin, but
this is an optional package that is not part of the standard Wagtail
installation. This commit adds a note to the README about this and also
adds a Django check that will fail if wagtail.contrib.modeladmin isn't
in INSTALLED_APPS.

As a side effect this change also adds a custom AppConfig for this app.