Add a GitHub Action to upload releases to PyPI
This commit adds our standard GitHub Action to upload releases to PyPI when the release is created in GitHub. This will remove the separate steps of building a wheel and sdist on the tag and uploading it manually when we create a release.

It will require the `TWINE_USERNAME` and `TWINE_PASSWORD` secrets to be created on this repository with the CFPB PyPI credentials. I am happy to do that.