Delete by migration accidentally included ShareableRoutablePage
This deletes the ShareableRoutablePage model that was accidentally included in #43 via migration. This should fix missing migration errors when users use 2.5.0.

I'm doing this via delete migration since the migration that added the model went out as part of 2.5.0. I don't want to leave migrations in a failed state for anyone.