Fix two Django warnings
This commit fixes two Django warnings that appear on startup:

?: (wagtailadmin.W003) The WAGTAILADMIN_BASE_URL setting is not defined
        HINT: This should be the base URL used to access the Wagtail admin site. Without this, URLs in notification emails will not display correctly.

As of Wagtail 3.0, we need to define the WAGTAILADMIN_BASE_URL setting:
https://docs.wagtail.org/en/stable/releases/3.0.html#base-url-setting-renamed-to-wagtailadmin-base-url

wagtailsharing.SharingSite: (models.W042) Auto-created primary key used when not defining a primary key type, by default 'django.db.models.AutoField'.
        HINT: Configure the DEFAULT_AUTO_FIELD setting or the WagtailSharingAppConfig.default_auto_field attribute to point to a subclass of AutoField, e.g. 'django.db.models.BigAutoField'.

To comply the Django 3.2 recommendation around use of BigAutoField for
model primary keys, a migration has been added that alters the type
of the PageBlock primary key to this type.
https://docs.djangoproject.com/en/4.1/releases/3.2/#customizing-type-of-auto-created-primary-keys