Add Sharing Link button to the page header
Wagtail 4 has moved the page header buttons into a drop-down menu, allowing more space for new buttons.

This change adds the “View sharing link” for any page to its header buttons menu, which appear on both the page’s listing view of child pages as well as its edit view.

This means that for every page, the “View sharing link” now appears in the “More” context menu in the parent’s listing view, in the header menu in its own listing view, and in the header menu in its edit view.