Fix: get_sharing_url should use database slug
Currently the get_sharing_url helper method can return an incorrect
sharing URL for pages whose slug has changed since it was published.
The sharing URL needs to use the slug associated with the version of
the page in the database, because that's the slug that gets used when
incoming requests are being routed.

This also impacts pages that have never been published but have
multiple draft versions with different slugs; the sharing URL needs to
use the initial slug because, as of the latest versions of Wagtail,
that is still the version that is used for slug-based routing.

This fix updates the get_sharing_url logic to refresh the page object
from the database before calculating the URL.

This addresses CFPB internal D&CP#266.