Deprecate reliance on Wagtail's testapp
Currently we use Wagtail's wagtail.test.testapp as part of our test
settings in order to use some of its test models, for example
SimplePage.

As reported in issue 66, there's a problem with the migrations for the
testapp in Wagtail 4.2.4 -- see wagtail#10724.

This commit adds a new simple TestPage to the testapp in this repo
which we can use in the tests to avoid relying on Wagtail's testapp.