Add support for Wagtail 5.0
This commit adds explicit support for Wagtail 5.0 (not 5.1 or 5.2 yet)
by adding it to tox.ini and the GitHub Actions tests.

Along with this change comes testing against Django 4.2.

In Wagtail 5.1, modeladmin begins to be deprecated in favor of Wagtail
snippets:

https://docs.wagtail.org/en/stable/releases/5.1.html#wagtail-contrib-modeladmin-is-deprecated

For this reason, it makes sense to release a 5.0-specific version and
then a subsequent 5.1+ version that uses snippets instead of modeladmin,
instead of trying to support both approaches simultaneously.