Support multiple settings-based sharing hosts
This change modifies the behavior of the WAGTAILSHARING_HOST setting,
when used with the SettingsHostRouter router. Instead of specifying a
single sharing host, it's now possible to specify multiple sharing
hosts, either as a list or as a single comma-delimited string.

This supports testing of wagtail-sharing where you may have multiple
domains, say, one for internal testing and one that is public-facing.